requirements.txt
requests==2.10.0
pycrypto==2.6.1
pytest==3.0.7
PyYAML==3.12
PyMySQL==0.7.9
pytest-html==1.14.2
